+++ spec/api_global_models_spec.rb
@@ -8,6 +8,10 @@ describe 'Global Models' do
         class Thing < Grape::Entity
           expose :text, documentation: { type: 'string', desc: 'Content of something.' }
         end
+
+        class CombinedThing < Grape::Entity
+          expose :text, documentation: { type: 'string', desc: 'Content of something.' }
+        end
       end
     end
   end
@@ -18,6 +22,14 @@ describe 'Global Models' do
       get '/thing' do
         thing = OpenStruct.new text: 'thing'
         present thing, with: Entities::Some::Thing
+      end
+
+      desc 'This gets combined thing.',
+           params: Entities::Some::CombinedThing.documentation,
+           entity: Entities::Some::CombinedThing
+      get '/combined_thing' do
+        thing = OpenStruct.new text: 'thing'
+        present thing, with: Entities::Some::CombinedThing
       end
 
       add_swagger_documentation models: [Entities::Some::Thing]
@@ -38,5 +50,27 @@ describe 'Global Models' do
             'text' => { 'type' => 'string', 'description' => 'Content of something.' }
           }
         })
+  end
+
+  it 'uses global models and route endpoint specific entities together' do
+    get '/swagger_doc/combined_thing.json'
+    json = JSON.parse(last_response.body)
+
+    expect(json['models']).to include(
+                                  'Some::Thing' => {
+                                    'id' => 'Some::Thing',
+                                    'properties' => {
+                                      'text' => { 'type' => 'string', 'description' => 'Content of something.' }
+                                    }
+                                  })
+
+    expect(json['models']).to include(
+                                  'Some::CombinedThing' => {
+                                    'id' => 'Some::CombinedThing',
+                                    'properties' => {
+                                      'text' => { 'type' => 'string', 'description' => 'Content of something.' }
+                                    }
+                                  })
+
   end
 end